Split SotreSessionProgress to combinators.
